fix: use node coordinates with sign

+++ L2jGeodataPathFinder/GeoData/Converter.h
@@ -11,8 +11,8 @@ public:
 	{
 		return Point
 		{
-			(uint32_t)std::floor((x / Constants::BLOCK_SIZE_IN_GAME_COORDS + Constants::DELTA_COL) * Constants::REGION_CELLS_DIMENSION),
-			(uint32_t)std::floor((y / Constants::BLOCK_SIZE_IN_GAME_COORDS + Constants::DELTA_ROW) * Constants::REGION_CELLS_DIMENSION)
+			(int32_t)std::floor((x / Constants::BLOCK_SIZE_IN_GAME_COORDS + Constants::DELTA_COL) * Constants::REGION_CELLS_DIMENSION),
+			(int32_t)std::floor((y / Constants::BLOCK_SIZE_IN_GAME_COORDS + Constants::DELTA_ROW) * Constants::REGION_CELLS_DIMENSION)
 		};
 	}
 
@@ -20,8 +20,8 @@ public:
 	{
 		return Point
 		{
-			(uint32_t)std::floor(coords.x / Constants::REGION_CELLS_DIMENSION),
-			(uint32_t)std::floor(coords.y / Constants::REGION_CELLS_DIMENSION)
+			(int32_t)std::floor(coords.x / Constants::REGION_CELLS_DIMENSION),
+			(int32_t)std::floor(coords.y / Constants::REGION_CELLS_DIMENSION)
 		};
 	}
 
@@ -29,8 +29,8 @@ public:
 	{
 		return Point
 		{
-			(coords.x - Constants::DELTA_COL * Constants::REGION_CELLS_DIMENSION) * Constants::CELL_SIZE_IN_GAME_COORDS,
-			(coords.y - Constants::DELTA_ROW * Constants::REGION_CELLS_DIMENSION) * Constants::CELL_SIZE_IN_GAME_COORDS,
+			(int32_t)(coords.x - Constants::DELTA_COL * Constants::REGION_CELLS_DIMENSION) * Constants::CELL_SIZE_IN_GAME_COORDS,
+			(int32_t)(coords.y - Constants::DELTA_ROW * Constants::REGION_CELLS_DIMENSION) * Constants::CELL_SIZE_IN_GAME_COORDS,
 		};
 	}
 };
